refactor(dispatch): isolate agent workflows
